Palm_Routing/requirements.txt
google-generativeai
PyPDF2
pymongo
docx2txt
flask